Creation of the topic registration, used to delimit the subjects discussed in the user group. Improvements on the mailing list registration, allowing the edition of a mailing list and the visualization of its subscribers. Fixing the upload of partners' logo. Changing once again the look and feel of the application to a more light appearance.

+++ jug-ejb/src/org/cejug/knowledge/business/TopicBsn.java
@@ -56,6 +56,9 @@ public class TopicBsn {
     }
 
     public void remove(String name) {
+        if(name == null || name.isEmpty())
+            return;
+        
         Topic topic = em.find(Topic.class, name);
         if(topic != null)
             em.remove(topic);
